adjust cursor position to be within the extended code block
